feat: make DataGrid responsive - render as cards on mobile view #19
